Redistributable packaging
Being able to install with pip and/or python setup.py is always a
nice thing to have. Plus, having a package that gets installed in
the python library makes it super easy to have reusable code in
modules that can be shared across plugins and other utilities.

The setup script uses pbr (python build reasonableness) which is
already successfully used by all OpenStack projects and makes
managing of requirements and versioning really easy.