Fix processing of OVN NB DB monitor notifications
The code, based on a simple split was not handling correctly lists
with multiple items, which are separated by a space in the monitor
output.
This patch introduces a regular expression to divide fields properly,
and adjusts the routine for converting external_ids into a dict, as
that routine was relying on a hack in parsing monitor output which
has now been removed.

Also, the routine from creating an ACL in OVN is now using shlex as
it turns out this simplify escaping for the match rule.