OVN watcher: correctly handle port creation and deletion
This patch ensures the ACLs are correctly configured when a logical
port is created. Due to the way the CNI plugin currently operates,
a row change even must be detected, to fetch the operation when
the pod name is added to the logical port external ids.

This patch also ensures ACLs are deleted when a container is removed,
adding appropriate logic for finding the logical switch where the
ACL should be removed, since a query by logical port is unlikely to
work upon a delete event.