Store northbound DB connection URI in local ovs external ids
After separating SB and NB ovsdb-server processes, the ovn_remote
external ids points to the SB DB only.
As the plugin needs to connect to the NB DB, this patch updates the
init process to accept the NB DB connection URI and stores it in
Open_vSwitch external ids.

The patch also factors out from the watcher oslo_config init.
However, the plugin does not use it as it does not play well with
argparse that we currently need for dispatching plugin commands.