Kubernetes and OVN event processing system
This patch implements a simple event processing system based on a
process-local priority queue. Events are detected by the following
watchers:
- k8s pod watcher
- k8s ns annotation watcher
- k8s network policy watcher
- OVN logical port monitor

The watchers perform some filtering and formatting on events and
then send them to a queue.
A policy processor retrieves events from this queue, and processes
them, with the goal of tranforming those into appropriate changes
in OVN ACLS.

This patch so far implements basic traffic whitelisting.