Process ACLs for whitelisted traffic
This patch adds ACL processing for pods whose traffic must be
whitelisted as they're deployed on namespaces with no isolation

The patch also restructures the startup logic in order to ensure
the policy processor has easy access to the watcher instance which
keep "caches" of the monitored objects.

When an object is not found in the cache, it is retrieved by the
API server. This behaviour is implemented in this patch for pods
and namespaces.