Load network policies only once when doing pod ACL
Loading policies for every pod can be very expensive.
This patch prefetches policies before doing pod ACLs.
Note that the algorithm could use the policy cache, but until we can
be reasonably sure the cache is always in sync or know when it's
dirty it's not the case to use it for list operations.