Implement both ports and from clause as OVN acls
This patch takes care of implementing both the 'ports' clause and
the 'from' clause. As for the latter, the pod selector, if specified
will be applied only to the current namespace, whereas the namespace
selector will imply that all pods in selected namespace will be
allowed as ingress traffic on the pods where the network policy
is applied.

As a part of this patch, several bugs in the previous code which
handled the ports selector have been fixed. In particular,
kubernetes label selector are now always builtin using set logic,
as it works fine even with a single element.