Regenerate watchers when k8s connection drops
Periodically the kubernetes server times out watch API requests.
These results in a StopIteration exception on the result stream.

This patch handles this exception and handles it regenerating
the watcher (and obviously sending another watch API request to
the server).

Watcher regeneration capability is also provided for OVN DB,
even if there is no timeout on monitoring there.