Make plugin independent from ovn-northd
This patch simplies the CNI plugin removing calls to the northbound
database. This considerably simplifies locking down access to the nb
database, and removes any need for a OVN watcher.

Operations can be summarized as follows:
- The plugin generates a MAC and IP address for the pod, and annotates
  the pod itself with these information.
- The pod watcher creates the OVN logical port when the following
  information are available: pod MAC, pod IP, pod Node, pod Infra
  Container ID. When the logical port is created, a default 'drop-all'
  is still added. Further, a policy event is generated so that the
  appropriate ACLs for the pod are created
- ACLs are now deleted when the pod DELETED event is captured and not
  anymore wehn the logical port DELETE event is captured.
- Kubernetes Network policies are still managed as before

The command 'init' has not been removed from the CNI plugin but it
now only initializes a few variables which are then needed by the
CNI plugin. These variables are stored in the Open_vSwitch table of
the OVS db. All the operations concerning the logical topology for
the kubernetes cluster have been removed to the plugin and moved
to a new utility, ovn_k8s_init_host

This utility accepts a subnet CIDR and a node name parameters as input,
ensures the presence of a logical switch for the node, and connects
said logical switch to a logical router where all the other nodes are
connected.